+++ CHANGES.md
@@ -4,7 +4,7 @@ Changes to mtools
 
 #### version 1.1.1
 
-  * mplotqueries: use start times of operations that have a duration, rather than end time with `--xpos start` (#130)
+  * mplotqueries: use start times of operations that have a duration, rather than end time with `--optime-start` (#130)
   * mplotqueries: group by query pattern with `--group pattern` (#129).
   * mlaunch: allow more than 7 nodes, everyone above 7 is non-voting (#123).
   * mloginfo: fixed bug where anonymous unix sockets can't be parsed (#121).